% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keccak256.R
\name{keccak256}
\alias{keccak256}
\title{Compute Keccak256 hash of a string}
\usage{
keccak256(inputs, hexConvert = TRUE)
}
\arguments{
\item{inputs}{A vector string or raw bytes to hash}

\item{hexConvert}{If the string starts with 0x, convert it to raw bytes. Default is TRUE.}
}
\value{
A character string containing the hexadecimal representation of the hash
}
\description{
This function computes the Keccak256 hash of a string and returns the result as a hexadecimal string.
}
\examples{
keccak256("")
keccak256("hello world")
keccak256("0x68656c6c6f20776f726c64")
keccak256("0x68656c6c6f20776f726c64", hexConvert = FALSE)
}
