% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keccak256.R
\name{keccak256_file}
\alias{keccak256_file}
\title{Compute Keccak256 hash of a file}
\usage{
keccak256_file(filepath)
}
\arguments{
\item{filepath}{Path to the file to hash}
}
\value{
A character string containing the hexadecimal representation of the hash
}
\description{
This function computes the Keccak256 hash of a file and returns the result as a hexadecimal string.
}
\examples{
\dontrun{
keccak256_file("path/to/file.txt")
}
}
