% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keccak256.R
\name{keccak256_integer}
\alias{keccak256_integer}
\title{Compute Keccak256 hash of an integer's bytes}
\usage{
keccak256_integer(integers, intBytes = 32)
}
\arguments{
\item{integers}{An integer vector}

\item{intBytes}{Number of bytes to represent the number (that will be hashed)}
}
\value{
A character string containing the hexadecimal representation of the hash
}
\description{
This function defaults to 32 bytes and will error if the number is not an
integer.
}
\examples{
keccak256_integer(0)
keccak256_integer(as.integer(0), intBytes = 1)
}
