% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keccak256r-package.R
\docType{package}
\name{keccak256r-package}
\alias{keccak256r}
\alias{keccak256r-package}
\title{keccak256r: The Keccak256 hash algorithm for R}
\description{
More about what it does (maybe more than one line). Continuation lines should be indented.
}
\author{
\strong{Maintainer}: Red Swan \email{redswan@outlook.com}

}
\keyword{internal}
